refactor: move rename utils location
