+++ src/shared/utils/rename.js
@@ -3,14 +3,14 @@ const PLUS = '+'
 const MINUS = '-'
 const OPERATORS = [PLUS, MINUS]
 
-export function getRuleString (out) {
+export const getRuleString = (out) => {
   const rule = out.match(RULE_REGEX)
   const result = rule && rule[1]
 
   return result
 }
 
-export function buildRule (rule) {
+export const buildRule = (rule) => {
   let ruleArr
   let operator = PLUS
   let init = 1
@@ -41,7 +41,7 @@ export function buildRule (rule) {
   }
 }
 
-export function buildOuts (uris = [], out = '') {
+export const buildOuts = (uris = [], out = '') => {
   const result = []
   const count = uris.length
   if (count === 0 || !out) {
